Fixup merging of CDATA handling with io::Write/buffer rework
The merger of the master branch (w/ the write_cdata_text() was not done
properly so the code does not compile anymore. This commit fixes things
to bring back the write_cdata_text() code (part of which was gone) while
moving from internal buffer writes to io::Write methods.

This commit also drops Clone and Debug (effectively reverting
9912b2ac79b58e8b1b28db092d0e515eebb19ab0) as I can't make it actually
work in tests to output to the same buffer/io::Write, as it clones the
underlying writer too. In any case, I don't see the point in having
Clone to write part of the XML in a different manner, as the only thing
that's public in that matter is preserve_whitespaces, which can just be
set to false/true around the code that you want to write differently,
rather than cloning the XmlWriter.